// src/app/theme/shared/components/toasts/toasts.enums.ts
export enum ToastType {
  Success = 'success',
  Error = 'error',
  Warning = 'warning',
  Info = 'info'
}

export enum ToastPosition {
  TopLeft = 'top-left',
  TopRight = 'top-right',
  TopCenter = 'top-center',
  BottomLeft = 'bottom-left',
  BottomRight = 'bottom-right',
  BottomCenter = 'bottom-center'
}

export enum ToastAnimation {
  SlideIn = 'slide-in',
  FadeIn = 'fade-in',
  SlideUp = 'slide-up',
  SlideDown = 'slide-down',
  Bounce = 'bounce',
  Scale = 'scale'
}